refactor: server.tool() を registerTool() に移行（MCP SDK 非推奨 API 対応）
@modelcontextprotocol/sdk v1.27.1 で deprecated となった server.tool() を
新しい server.registerTool() API に一括移行。

- 全17ファイルの server.tool(name, desc, schema, cb) を
  server.registerTool(name, { description, inputSchema }, cb) に変換
- wrapServerWithMetrics の Proxy を registerTool 対応に更新
- jobs.ts の max-lines 超過を解消するため smelt 関連を smelting.ts に分離
- smelt.spec.ts のインポートパスとモック形式を更新
