+++ packages/mcp/src/tools/mc-memory.ts
@@ -68,22 +68,29 @@ export function registerMcMemoryTools(server: McpServer, deps: McMemoryDeps): vo
 
 	// --- Goals ---
 
-	server.tool("mc_read_goals", "Minecraft 目標ファイルを読む（現在の目標のみ）", {}, () => {
-		const content = readOverlay(dataDir, GOALS_FILENAME);
-		return {
-			content: [{ type: "text" as const, text: content || "(目標ファイルは空です)" }],
-		};
-	});
+	server.registerTool(
+		"mc_read_goals",
+		{ description: "Minecraft 目標ファイルを読む（現在の目標のみ）" },
+		() => {
+			const content = readOverlay(dataDir, GOALS_FILENAME);
+			return {
+				content: [{ type: "text" as const, text: content || "(目標ファイルは空です)" }],
+			};
+		},
+	);
 
-	server.tool(
+	server.registerTool(
 		"mc_update_goals",
-		"Minecraft 目標ファイルを上書き更新する（バックアップ自動作成）。現在の目標のみ記載すること。達成済み目標や探索メモは mc_update_progress に記録する。",
 		{
-			content: z
-				.string()
-				.min(1)
-				.max(MAX_GOALS_CHARS)
-				.describe("新しい MINECRAFT-GOALS.md の内容（最大 20,000 文字）"),
+			description:
+				"Minecraft 目標ファイルを上書き更新する（バックアップ自動作成）。現在の目標のみ記載すること。達成済み目標や探索メモは mc_update_progress に記録する。",
+			inputSchema: {
+				content: z
+					.string()
+					.min(1)
+					.max(MAX_GOALS_CHARS)
+					.describe("新しい MINECRAFT-GOALS.md の内容（最大 20,000 文字）"),
+			},
 		},
 		({ content }) => {
 			writeOverlay(dataDir, GOALS_FILENAME, content);
@@ -100,10 +107,12 @@ export function registerMcMemoryTools(server: McpServer, deps: McMemoryDeps): vo
 
 	// --- Progress ---
 
-	server.tool(
+	server.registerTool(
 		"mc_read_progress",
-		"Minecraft ワールド進捗を読む（装備段階、拠点、探索範囲、主要資源、達成済み目標、プレイヤーメモ）",
-		{},
+		{
+			description:
+				"Minecraft ワールド進捗を読む（装備段階、拠点、探索範囲、主要資源、達成済み目標、プレイヤーメモ）",
+		},
 		() => {
 			const content = readOverlay(dataDir, PROGRESS_FILENAME);
 			return {
@@ -112,15 +121,18 @@ export function registerMcMemoryTools(server: McpServer, deps: McMemoryDeps): vo
 		},
 	);
 
-	server.tool(
+	server.registerTool(
 		"mc_update_progress",
-		"Minecraft ワールド進捗を更新する（バックアップ自動作成）。装備段階、拠点、探索範囲、主要資源、達成済み目標、プレイヤーメモを記録する。",
 		{
-			content: z
-				.string()
-				.min(1)
-				.max(MAX_PROGRESS_CHARS)
-				.describe("新しい MINECRAFT-PROGRESS.md の内容（最大 20,000 文字）"),
+			description:
+				"Minecraft ワールド進捗を更新する（バックアップ自動作成）。装備段階、拠点、探索範囲、主要資源、達成済み目標、プレイヤーメモを記録する。",
+			inputSchema: {
+				content: z
+					.string()
+					.min(1)
+					.max(MAX_PROGRESS_CHARS)
+					.describe("新しい MINECRAFT-PROGRESS.md の内容（最大 20,000 文字）"),
+			},
 		},
 		({ content }) => {
 			writeOverlay(dataDir, PROGRESS_FILENAME, content);
@@ -137,29 +149,32 @@ export function registerMcMemoryTools(server: McpServer, deps: McMemoryDeps): vo
 
 	// --- Skills ---
 
-	server.tool("mc_read_skills", "Minecraft スキルライブラリを読む", {}, () => {
+	server.registerTool("mc_read_skills", { description: "Minecraft スキルライブラリを読む" }, () => {
 		const content = readOverlay(dataDir, SKILLS_FILENAME);
 		return {
 			content: [{ type: "text" as const, text: content || "(スキルライブラリは空です)" }],
 		};
 	});
 
-	server.tool(
+	server.registerTool(
 		"mc_record_skill",
-		"Minecraft スキルライブラリにスキルを追記する（有効条件・前提装備・失敗パターン付き）",
 		{
-			name: z.string().min(1).max(200).describe("スキル名"),
-			description: z.string().min(1).max(2_000).describe("スキルの説明・手順"),
-			preconditions: z
-				.string()
-				.max(500)
-				.optional()
-				.describe("有効条件・前提装備（例: 石のピッケル以上が必要）"),
-			failure_patterns: z
-				.string()
-				.max(500)
-				.optional()
-				.describe("既知の失敗パターン（例: 夜間は敵mobで中断されやすい）"),
+			description:
+				"Minecraft スキルライブラリにスキルを追記する（有効条件・前提装備・失敗パターン付き）",
+			inputSchema: {
+				name: z.string().min(1).max(200).describe("スキル名"),
+				description: z.string().min(1).max(2_000).describe("スキルの説明・手順"),
+				preconditions: z
+					.string()
+					.max(500)
+					.optional()
+					.describe("有効条件・前提装備（例: 石のピッケル以上が必要）"),
+				failure_patterns: z
+					.string()
+					.max(500)
+					.optional()
+					.describe("既知の失敗パターン（例: 夜間は敵mobで中断されやすい）"),
+			},
 		},
 		({ name, description, preconditions, failure_patterns }) => {
 			const existing = readOverlay(dataDir, SKILLS_FILENAME);
