+++ packages/minecraft/src/actions/index.ts
@@ -4,7 +4,7 @@ import type { Logger } from "@vicissitude/shared/types";
 import type { JobManager } from "../job-manager.ts";
 import { registerAttackEntity } from "./combat.ts";
 import { registerSendChat, registerEquipItem, registerPlaceBlock } from "./interaction.ts";
-import { registerCraftItem, registerSleepInBed, registerSmeltItem } from "./jobs.ts";
+import { registerCraftItem, registerSleepInBed } from "./jobs.ts";
 import {
 	registerFollowPlayer,
 	registerGoTo,
@@ -12,6 +12,7 @@ import {
 	registerStop,
 } from "./movement.ts";
 import type { GetBot } from "./shared.ts";
+import { registerSmeltItem } from "./smelting.ts";
 import { registerSurvivalTools } from "./survival/index.ts";
 
 export function registerActionTools(
