+++ packages/minecraft/src/mcp-tools.ts
@@ -21,10 +21,9 @@ function registerObserveStateTool(
 	ctx: BotContext,
 	jobManager: JobManager,
 ): void {
-	server.tool(
+	server.registerTool(
 		"observe_state",
-		"Minecraft ボットの現在の状態を自然言語要約で取得する",
-		{},
+		{ description: "Minecraft ボットの現在の状態を自然言語要約で取得する" },
 		async () => {
 			const bot = ctx.getBot();
 			if (!bot || !bot.entity) {
@@ -69,20 +68,22 @@ function registerObserveStateTool(
 }
 
 function registerRecentEventsTool(server: McpServer, ctx: BotContext): void {
-	server.tool(
+	server.registerTool(
 		"get_recent_events",
-		"Minecraft ボットの直近イベントログをテキスト形式で取得する",
 		{
-			limit: z
-				.number()
-				.min(1)
-				.max(50)
-				.default(10)
-				.describe("取得するイベント数（デフォルト: 10、最大: 50）"),
-			importance: z
-				.enum(["low", "medium", "high"])
-				.optional()
-				.describe("最低重要度フィルタ（例: medium → medium 以上のみ）"),
+			description: "Minecraft ボットの直近イベントログをテキスト形式で取得する",
+			inputSchema: {
+				limit: z
+					.number()
+					.min(1)
+					.max(50)
+					.default(10)
+					.describe("取得するイベント数（デフォルト: 10、最大: 50）"),
+				importance: z
+					.enum(["low", "medium", "high"])
+					.optional()
+					.describe("最低重要度フィルタ（例: medium → medium 以上のみ）"),
+			},
 		},
 		({ limit, importance }) => {
 			const events = ctx.getEvents();
@@ -98,16 +99,18 @@ function registerRecentEventsTool(server: McpServer, ctx: BotContext): void {
 }
 
 function registerJobStatusTool(server: McpServer, jobManager: JobManager): void {
-	server.tool(
+	server.registerTool(
 		"get_job_status",
-		"現在のジョブ状態と直近のジョブ履歴を取得する",
 		{
-			limit: z
-				.number()
-				.min(1)
-				.max(20)
-				.default(5)
-				.describe("取得するジョブ履歴数（デフォルト: 5、最大: 20）"),
+			description: "現在のジョブ状態と直近のジョブ履歴を取得する",
+			inputSchema: {
+				limit: z
+					.number()
+					.min(1)
+					.max(20)
+					.default(5)
+					.describe("取得するジョブ履歴数（デフォルト: 5、最大: 20）"),
+			},
 		},
 		({ limit }) => {
 			const current = jobManager.getCurrentJob();
@@ -119,43 +122,42 @@ function registerJobStatusTool(server: McpServer, jobManager: JobManager): void 
 }
 
 function registerViewerUrlTool(server: McpServer, ctx: BotContext, viewerPort: number): void {
-	server.tool("get_viewer_url", "Minecraft ビューアーの URL を返す", {}, () => {
-		const bot = ctx.getBot();
-		if (!bot?.entity) {
-			return { content: [{ type: "text" as const, text: "ボット未接続" }] };
-		}
-		return {
-			content: [
-				{
-					type: "text" as const,
-					text: `http://localhost:${String(viewerPort)}`,
-				},
-			],
-		};
-	});
+	server.registerTool(
+		"get_viewer_url",
+		{ description: "Minecraft ビューアーの URL を返す" },
+		() => {
+			const bot = ctx.getBot();
+			if (!bot?.entity) {
+				return { content: [{ type: "text" as const, text: "ボット未接続" }] };
+			}
+			return {
+				content: [
+					{
+						type: "text" as const,
+						text: `http://localhost:${String(viewerPort)}`,
+					},
+				],
+			};
+		},
+	);
 }
 
 /**
- * server.tool() 呼び出しをインターセプトし、各ツールのハンドラ実行時にメトリクスを記録する
+ * server.registerTool() 呼び出しをインターセプトし、各ツールのハンドラ実行時にメトリクスを記録する
  * Proxy を使って McpServer を薄くラップすることで、個々のツール登録関数を変更せずに全ツールを計測できる
  */
 function wrapServerWithMetrics(server: McpServer, metrics: MetricsCollector): McpServer {
 	return new Proxy(server, {
 		get(target, prop, receiver) {
-			if (prop !== "tool") return Reflect.get(target, prop, receiver);
-			// oxlint-disable-next-line no-explicit-any -- McpServer.tool() は複数オーバーロードを持つため any で受ける
-			return (name: string, ...args: any[]) => {
-				const lastIdx = args.length - 1;
-				const originalHandler = args[lastIdx];
-				if (typeof originalHandler === "function") {
-					// oxlint-disable-next-line no-explicit-any -- handler の引数型はオーバーロードごとに異なる
-					args[lastIdx] = (...handlerArgs: any[]) => {
-						metrics.incrementCounter(METRIC.MC_MCP_TOOL_CALLS, { tool: name });
-						return originalHandler(...handlerArgs);
-					};
-				}
-				// oxlint-disable-next-line no-unsafe-function-type, ban-types -- target.tool の型を正確に表現できないため
-				return (target.tool as (...a: unknown[]) => unknown).call(target, name, ...args);
+			if (prop !== "registerTool") return Reflect.get(target, prop, receiver);
+			// oxlint-disable-next-line no-explicit-any -- McpServer.registerTool() のコールバック型を正確に表現できないため any で受ける
+			return (name: string, config: any, cb: (...handlerArgs: any[]) => any) => {
+				// oxlint-disable-next-line no-explicit-any -- handler の引数型はツールごとに異なる
+				const wrappedCb = (...handlerArgs: any[]) => {
+					metrics.incrementCounter(METRIC.MC_MCP_TOOL_CALLS, { tool: name });
+					return cb(...handlerArgs);
+				};
+				return target.registerTool(name, config, wrappedCb);
 			};
 		},
 	});
